Attribute all light modules; consolidate driver docs; CodeRabbit batch
Credit every effect, modifier, and layout with its original author and source (from MoonLight) in both the code and the docs, consolidate the six per-driver doc pages into one drivers.md the way effects.md works, and process a CodeRabbit review batch (fast-board plasma stall, praxis overflow, blurz radius edge, present-tense docs, order-independent img parsing, plus new regression tests).

KPI: 16384lights | PC:465KB | tick:256/191/1/250/40/4/533/120/45/37/314/248/42/2/68us(FPS:3906/5235/1000000/4000/25000/250000/1876/8333/22222/27027/3184/4032/23809/500000/14705) | tick:5116us(FPS:195) | heap:31522KB | src:166(30973) | test:90(13362) | lizard:113w

Core:
- AudioModule: reverted a CodeRabbit suggestion to gate the active_ seat on inited_ — a claimed-but-uninited module publishes valid silence (the documented latestFrame() contract, pinned by the two-mic test), so the first-wins model is correct; the change would have broken the desktop/no-mic case.

Light domain:
- Attribution: every effect/modifier/layout .h carries an `// Author: <name(s)> — <MoonLight source URL>` line in its header comment (Andrew Tuline / WLED-SR, @softhack007, limpkin, ldirko & blazoncek, @Brandon502 / WildCats08, MONSOONO / @Flavourdynamics, @TroyHacks, Ewoud Wijma, Mark Kriegsman / FastLED, netmindz; project-level credit where MoonLight names no individual; "projectMM original" for native effects).
- PlasmaEffect: phase_ accumulates the raw numerator (dt·bpm·256) and the /60000 divide runs only at the sampling point — dividing per tick truncated sub-unit progress to zero on a fast board (short dt), stalling the animation.
- PraxisEffect: promote now·speed to 64-bit before the divide in huebase — it overflowed uint32 after a few hours' uptime and made the hue jump.
- BlurzEffect: dot radius stays 0 for a grid up to 32 (was 1 at exactly 32, contradicting the "≤32 = one pixel" contract); present-tense default comment.
- draw.h: box-blur comment corrected — a 1×N 1D layer blurs along Y (its only axis with extent>1), not X.

UI:
- (docs only this cycle — see Docs / CI)

Scripts / MoonDeck:
- moondeck.py: <img> parsing is now attribute-order-independent (a shared _img_attrs helper for the table-cell path + an inline extractor for the standalone-line path), so `<img src alt width>` in any order keeps every attribute instead of silently dropping the out-of-order ones.
- check_specs.py: added "Driver" to CONSOLIDATED_PAGES so the six leaf *Driver control names route to the new drivers.md (the Drivers container and the ParallelLedDriver CRTP base are excluded).

Tests:
- unit_Layer_persistence: added a rebuild-clear regression — a lit pixel must not survive Layer::onBuildState() (persistence holds between frames, not across rebuilds).
- unit_RainbowEffect: pin Palettes::setActive(0) before the non-black assertion so it no longer depends on prior tests' global palette state.
- unit_draw: present-tense comments (dropped "previously untested" / "regression here"); regenerated unit-tests.md and scenario-tests.md.

Docs / CI:
- drivers.md (new): the six drivers as one page like effects.md — description + controls + Source + a link to each driver's detail page; grouped LED output / Network / Smart light / Preview.
- The six detail pages (Rmt/Lcd/Parlio/NetworkSend/Hue/Preview) keep their rich reference material (wire contracts, buffer slicing, memory sizing, the RMT flicker-troubleshooting playbook, loopback self-tests, the P4 pin budget, per-protocol chunking, Hue HTTP contract, Preview binary protocol) with a backlink header; the moved description+controls are dropped.
- Attribution: every effect/modifier/layout .md entry gains an `Origin: <project> · by <author> · via [MoonLight] · source [<Name>.h]` line under its controls; the effects Source section completed to all 39.
- architecture.md: persistence section states it holds between frames, not across rebuilds (onBuildState clears); trimmed the D1 paragraph's layout-orientation over-reach.

Reviews:
- 🐇 Plasma fast-board stall (accumulator fix), Praxis 32-bit overflow (64-bit), Blurz radius ≤32 edge, draw.h 1D-blur comment, architecture.md persistence-across-rebuild + D1 wording, unit_draw present-tense, unit_Layer_persistence rebuild-clear test, unit_RainbowEffect palette pin, moondeck.py order-independent <img> — fixed.
- 🐇 AudioModule active_ ownership (silence-frame contract; would break desktop no-mic), SolidEffect sqrtf (3/frame cold path, not per-light), StarField float (≤255 stars off the per-pixel path, fidelity-locked), DemoReel swapTo null-after-failed-create (loop()'s !current_ guard already prevents the crash), scenario_Layer_base_pipeline.json tight contract (the 200us pc- tolerance floor already absorbs it; loosening hides regressions) — verified not valid against current code, skipped with reason.

+++ src/light/effects/PraxisEffect.h
@@ -25,6 +25,7 @@ namespace mm {
 // written fresh on EffectBase + the shared draw / palette / math8 primitives. Our beatsin16
 // takes the current time (elapsed()) as its second argument, matching the lib8tion shape
 // (bpm, timebase) with the time source threaded in at the domain edge.
+// Author: MONSOONO / @Flavourdynamics (MoonLight) — https://github.com/MoonModules/MoonLight/blob/main/src/MoonLight/Nodes/Effects/E_MoonLight.h
 class PraxisEffect : public EffectBase {
 public:
     const char* tags() const override { return "💫"; }  // MoonLight origin
@@ -73,7 +74,9 @@ public:
         const uint16_t micro = beatsin16(microMutatorFreq, now,
                                          microMutatorMin, microMutatorMax);
 
-        const uint32_t huebase = now * speed / 320;   // speed 8 == MoonLight's now/40; lower = calmer
+        // 64-bit before the multiply: now (millis) · speed overflows uint32 after a few hours' uptime,
+        // which would make the hue jump. speed 8 == MoonLight's now/40; lower = calmer.
+        const uint32_t huebase = static_cast<uint32_t>(static_cast<uint64_t>(now) * speed / 320);
         const int64_t  microDiv = static_cast<int64_t>(micro) + 1;  // micro+1, guards a divide-by-zero
 
         // hue = huebase + (x + y·macro·x) / (micro+1), truncated to the 0..255 palette wheel index.
